Assign environment variables from a .env file

This function reads environment variables from a specified .env file and assigns them to variables in the R environment based on a data frame containing variable names, values, and checks for directories and files. It is designed to facilitate the management of environment variables in a structured and reproducible manner.

Usage

assign_env_vars(env_file = ".env", env_variables_data = NULL)

Arguments

env_file

Character. Path to the environment file containing paths to data sources. Defaults to .env.

env_variables_data

data.frame. A data frame or tibble containing the columns var_name, value, check_dir, and check_file. Each row specifies an environment variable, the name to assign it to, and whether to check if it is a directory or file. This structure allows for additional validation checks on the variables being imported.

Value

This function is used for its side effects of setting environment variables in the R environment. It assigns each variable from the .env file to the R environment with the name specified in the env_variables_data data frame.

Author

Ahmed El-Gabbas