Check function arguments for specific types

This function checks if the specified arguments of a function match the expected type. It is useful for validating function inputs.

Usage

check_args(args_all, args_to_check, args_type)

Arguments

args_all

Character vector. Input parameters of the function. Usually as a result of formals() function

args_to_check

Character vector. Names of the arguments to be checked.

args_type

Character. The expected type of the arguments. Must be one of "character", "logical", or "numeric".

Value

The function does not return a value but will stop execution and throw an error if any of the specified arguments do not match the expected type.

Author

Ahmed El-Gabbas