Check if a raster stack reads from disk or memory

This function checks whether the layers of a RasterStack object are stored in memory or read from disk. It prints messages indicating whether all layers are in memory, all layers are on disk, or a mix of both. If there's a mix, it specifies which layers are on disk.

Usage

check_stack_in_memory(stack = NULL)

Arguments

stack

A RasterStack object. If NULL or not a RasterStack, the function will stop with an error.

Value

No return value, but prints messages to the console.

Author

Ahmed El-Gabbas

Examples

library(raster)
logo <- raster(system.file("external/rlogo.grd", package = "raster"))
logo@data@inmemory
#> [1] FALSE
logo@data@fromdisk
#> [1] TRUE
logo@file@name
#> [1] "/home/runner/work/_temp/Library/raster/external/rlogo.grd"

# -------------------------------------------

# A raster stack reading from files
ST2 <- raster::stack(logo, logo)
check_stack_in_memory(ST2)
#> All stack layers reads from memory
c(ST2[[1]]@data@inmemory, ST2[[2]]@data@inmemory)
#> [1] FALSE FALSE
c(ST2[[1]]@data@fromdisk, ST2[[2]]@data@fromdisk)
#> [1] TRUE TRUE
c(ST2[[1]]@file@name, ST2[[2]]@file@name)
#> [1] "/home/runner/work/_temp/Library/raster/external/rlogo.grd"
#> [2] "/home/runner/work/_temp/Library/raster/external/rlogo.grd"

# -------------------------------------------

logo2 <- raster::readAll(logo)
ST3 <- raster::stack(logo, logo2)
check_stack_in_memory(ST3)
#> [1] "Layers numbered (1) reads from disk"
c(ST3[[1]]@data@inmemory, ST3[[2]]@data@inmemory)
#> [1] FALSE  TRUE
c(ST3[[1]]@data@fromdisk, ST3[[2]]@data@fromdisk)
#> [1]  TRUE FALSE
c(ST3[[1]]@file@name, ST3[[2]]@file@name)
#> [1] "/home/runner/work/_temp/Library/raster/external/rlogo.grd"
#> [2] ""