Set up or stop parallel processing plan

Configures parallel processing with future::plan() or stops an existing plan. When stopping, it resets to sequential mode.

Usage

set_parallel(
  n_cores = 1L,
  strategy = "future::multisession",
  stop_cluster = FALSE,
  show_log = TRUE,
  future_max_size = 500L,
  ...
)

Arguments

n_cores

Integer. Number of cores to use. If NULL, defaults to sequential mode. Default is 1.

strategy

Character. The parallel processing strategy to use. Valid options are future::sequential (sequential), future::multisession (default), future::multicore (not supported on Windows), and future::cluster. If strategy is not one of the valid options or if future::multicore on Windows PC, it defaults to future::multisession. See future::plan() for more details.

stop_cluster

Logical. If TRUE, stops any parallel cluster and resets to sequential mode. If FALSE (default), sets up a new plan.

show_log

Logical. If TRUE (default), logs messages via cat_time().

future_max_size

Numeric. Maximum allowed total size (in megabytes) of global variables identified. See future.globals.maxSize argument of future::future.options for more details. Default is 500L for 500 MB.

...

Additional arguments to pass to cat_time.

Author

Ahmed El-Gabbas

Examples

# Prepare working in parallel
ecokit::set_parallel(n_cores = 2)
#> Setting up parallel processing with 2 core(s) - 02:22:07
future::plan()
#> multisession:
#> - args: function (..., workers = c(system = 2L), envir = parent.frame())
#> - tweaked: TRUE
#> - call: future::plan(strategy = strategy, workers = n_cores)
#> MultisessionFutureBackend:
#> Inherits: ClusterFutureBackend, MultiprocessFutureBackend, FutureBackend
#> Number of workers: 2
#> Number of free workers: 2
#> Available cores: 4
#> Automatic garbage collection: FALSE
#> Early signaling: TRUE
#> Interrupts are enabled: TRUE
#> Maximum total size of globals: 500.00 MiB
#> Maximum total size of value: 500.00 MiB
#> Argument ‘persistent’: FALSE
#> Argument ‘wait.timeout’: 86400
#> Argument ‘wait.interval’: 0.01
#> Argument ‘wait.alpha’: 1.01
#> Argument ‘hooks’: FALSE
#> Number of active futures: 0
#> Number of futures since start: 0
#> Workers of type RichSOCKcluster:
#> - Summary: Socket cluster with 2 nodes where 2 nodes are on host ‘localhost’ (R version 4.5.0 (2025-04-11), platform x86_64-pc-linux-gnu)
#> - [ OK ] Node 1/2: valid connection, alive [RichSOCKnode of a socket cluster on local host 'localhost' with pid 9420 (R version 4.5.0 (2025-04-11), x86_64-pc-linux-gnu) using socket connection #5 ('<-localhost:11388')]
#> - [ OK ] Node 2/2: valid connection, alive [RichSOCKnode of a socket cluster on local host 'localhost' with pid 9419 (R version 4.5.0 (2025-04-11), x86_64-pc-linux-gnu) using socket connection #6 ('<-localhost:11388')]

# ---------------------------------------------

# Stopping parallel processing
ecokit::set_parallel(stop_cluster = TRUE)
#> Stopping parallel processing - 02:22:08
future::plan()
#> sequential:
#> - args: function (..., gc = TRUE, envir = parent.frame(), workers = "<NULL>")
#> - tweaked: TRUE
#> - call: future::plan("future::sequential", gc = TRUE)
#> FutureBackend to be launched

# ---------------------------------------------

# Prepare working in parallel using `future::cluster`
ecokit::set_parallel(n_cores = 2, strategy = "future::cluster")
#> Setting up parallel processing with 2 core(s) - 02:22:08
future::plan()
#> cluster:
#> - args: function (..., workers = c(system = 2L), envir = parent.frame())
#> - tweaked: TRUE
#> - call: future::plan(strategy = strategy, workers = n_cores)
#> ClusterFutureBackend:
#> Inherits: MultiprocessFutureBackend, FutureBackend
#> Number of workers: 2
#> Number of free workers: 2
#> Available cores: 4
#> Automatic garbage collection: FALSE
#> Early signaling: TRUE
#> Interrupts are enabled: FALSE
#> Maximum total size of globals: 500.00 MiB
#> Maximum total size of value: 500.00 MiB
#> Argument ‘persistent’: FALSE
#> Argument ‘wait.timeout’: 86400
#> Argument ‘wait.interval’: 0.01
#> Argument ‘wait.alpha’: 1.01
#> Argument ‘hooks’: FALSE
#> Number of active futures: 0
#> Number of futures since start: 0
#> Workers of type RichSOCKcluster:
#> - Summary: Socket cluster with 2 nodes where 2 nodes are on host ‘localhost’ (R version 4.5.0 (2025-04-11), platform x86_64-pc-linux-gnu)
#> - [ OK ] Node 1/2: valid connection [RichSOCKnode of a socket cluster on local host 'localhost' with pid 9491 (R version 4.5.0 (2025-04-11), x86_64-pc-linux-gnu) using socket connection #5 ('<-localhost:11882')]
#> - [ OK ] Node 2/2: valid connection [RichSOCKnode of a socket cluster on local host 'localhost' with pid 9490 (R version 4.5.0 (2025-04-11), x86_64-pc-linux-gnu) using socket connection #6 ('<-localhost:11882')]

# Stopping parallel processing
ecokit::set_parallel(stop_cluster = TRUE)
#> Stopping parallel processing - 02:22:08
future::plan()
#> sequential:
#> - args: function (..., gc = TRUE, envir = parent.frame(), workers = "<NULL>")
#> - tweaked: TRUE
#> - call: future::plan("future::sequential", gc = TRUE)
#> FutureBackend to be launched