List and Download Files from a Zenodo Record

These functions provide an R interface to list all files in a Zenodo record and to download one of those files (optionally reading it into R).

Usage

zenodo_file_list(record_id)

download_zenodo_file(
  record_id = NULL,
  zenodo_file = NULL,
  dest_file = NULL,
  read_func = NULL,
  ...
)

Arguments

record_id

Character or numeric. The Zenodo record ID (e.g., "1234567").

zenodo_file

Character. The file key (or filename) as listed in the Zenodo record.

dest_file

Character (optional). Local destination path to save the downloaded file. If NULL (default), a temporary file is created and deleted if read_func is used.

read_func

Function (optional). A function to read or process the downloaded file (e.g., read.csv(), readr::read_csv(), readLines(), readxl::read_excel(), etc.). If NULL, only the file path is returned.

...

Additional arguments passed to read_func.

Value

zenodo_file_list returns a tibble with metadata for each file in the Zenodo record, including file key, filename, download link, and record-level metadata such as title, DOI, and creation dates.

download_zenodo_file: If read_func is NULL, returns the path to the downloaded file. If read_func is provided, returns the output of that function.

Details

zenodo_file_list() lists all files and their metadata for a given Zenodo record.

download_zenodo_file() downloads a specified file from a Zenodo record and can optionally read its contents into R using a user-supplied function.

Author

Ahmed El-Gabbas

Examples

require(ecokit)
require(terra)

files <- zenodo_file_list("6620748")
dplyr::glimpse(files)
#> Rows: 59
#> Columns: 12
#> $ title       <chr> "Phanerozoic continental climate and Köppen–Geiger climate…
#> $ id_url      <chr> "https://zenodo.org/api/records/6620748", "https://zenodo.…
#> $ created     <dttm> 2022-06-07 15:21:05, 2022-06-07 15:21:05, 2022-06-07 15:2…
#> $ modified    <dttm> 2024-07-16 15:44:29, 2024-07-16 15:44:29, 2024-07-16 15:4…
#> $ updated     <dttm> 2024-07-16 15:44:29, 2024-07-16 15:44:29, 2024-07-16 15:4…
#> $ doi_version <chr> "10.5281/zenodo.6620748", "10.5281/zenodo.6620748", "10.52…
#> $ doi         <chr> "10.5281/zenodo.6402040", "10.5281/zenodo.6402040", "10.52…
#> $ id          <int> 6620748, 6620748, 6620748, 6620748, 6620748, 6620748, 6620…
#> $ key         <chr> "40Ma_Pohletal2022_DIB_PhaneroContinentalClimate.nc", "0Ma…
#> $ size        <int> 8261200, 14498683, 8261200, 14573851, 8261200, 14407735, 8…
#> $ checksum    <chr> "md5:e9ca6e5e58b7b892c2dfd00fffdcc7af", "md5:65ee19cbf3da6…
#> $ link        <chr> "https://zenodo.org/api/records/6620748/files/40Ma_Pohleta…

# --------------------------------------------

# Download as file only
pdf_file <- download_zenodo_file(
   record_id = "1234567", zenodo_file = "article.pdf")
print(pdf_file)
#> /tmp/Rtmp4pYJ2G/article246f1d489daa.pdf

ecokit::file_type(pdf_file)
#> [1] "PDF document, version 1.6"
fs::file_delete(pdf_file)

# Download and read as data frame
nc_file <- download_zenodo_file(
  record_id = "6620748",
  zenodo_file = "100Ma_Pohletal2022_DIB_PhaneroContinentalClimate.nc",
  read_func = terra::rast)

print(class(nc_file))
#> [1] "SpatRaster"
#> attr(,"package")
#> [1] "terra"

print(nc_file)
#> class       : SpatRaster 
#> size        : 128, 128, 3  (nrow, ncol, nlyr)
#> resolution  : 1.40625, 2.8125  (x, y)
#> extent      : -90, 90, 0, 360  (xmin, xmax, ymin, ymax)
#> coord. ref. :  
#> sources     : 100Ma_Pohletal2022_DIB_PhaneroContinentalClimate246f4c4a3fc.nc:topo  
#>               100Ma_Pohletal2022_DIB_PhaneroContinentalClimate246f4c4a3fc.nc:koppen  
#>               100Ma_Pohletal2022_DIB_PhaneroContinentalClimate246f4c4a3fc.nc:area  
#> varnames    : topo (topography) 
#>               koppen (Koppen-Geiger climatic zones) 
#>               area (grid point area) 
#> names       :              topo,                                              koppen, area 
#> unit        : m above sea level, Koppen_Geiger classes (see associated data article),   m2 
fs::file_delete(nc_file)
#> Error: [is.na] cannot read from NETCDF:"/tmp/Rtmp4pYJ2G/100Ma_Pohletal2022_DIB_PhaneroContinentalClimate246f4c4a3fc.nc":topo