Check function arguments for specific types

This function checks if the specified arguments of a function match the expected type. It is useful for validating function inputs.

Usage

check_args(args_all, args_to_check, args_type)

Arguments

args_all

Character vector. Input parameters of the function. Usually as a result of formals() function

args_to_check

Character vector. Names of the arguments to be checked.

args_type

Character. The expected type of the arguments. Must be one of "character", "logical", or "numeric".

Value

The function does not return a value but will stop execution and throw an error if any of the specified arguments do not match the expected type.

Author

Ahmed El-Gabbas

Examples

f1 <- function(x = "AA", y = "BB", z = 1) {
  all_arguments <- ls(envir = environment())
  all_arguments <- purrr::map(
    all_arguments,
    function(x) get(x, envir = parent.env(env = environment()))) %>%
    stats::setNames(all_arguments)

 # Check if x and y are a character
 check_args(
    args_all = all_arguments, args_type = "character",
    args_to_check = c("x", "y"))

 # Check if z is a numeric
 check_args(
    args_all = all_arguments, args_type = "numeric",
    args_to_check = "z")

 # the rest of the function
 }

 # no output as x is a character
 f1(x = "X")

 # no output as z is a numeric
 f1(z = 20)

 # error as x is not a character
 try(f1(x = 1))
#> Error in check_args(args_all = all_arguments, args_type = "character",  : 
#>   The following character argument(s) must be character
#>   >>  x - 18:29:28
#> 
#> ----- Metadata -----
#> 
#> length_MissingArgs [length(missing_arguments)]: <integer>
#> 1
#> 
#> ----- Backtrace -----
#>   1. └─pkgdown::build_site_github_pages(new_process = FALSE, install = FALSE)
#>   2.   └─pkgdown::build_site(...)
#>   3.     └─pkgdown:::build_site_local(...)
#>   4.       └─pkgdown::build_reference(...)
#>   5.         ├─pkgdown:::unwrap_purrr_error(...)
#>   6.         │ └─base::withCallingHandlers(...)
#>   7.         └─purrr::map(...)
#>   8.           └─purrr:::map_("list", .x, .f, ..., .progress = .progress)
#>   9.             ├─purrr:::with_indexed_errors(...)
#>  10.             │ └─base::withCallingHandlers(...)
#>  11.             ├─purrr:::call_with_cleanup(...)
#>  12.             └─pkgdown (local) .f(.x[[i]], ...)
#>  13.               ├─base::withCallingHandlers(...)
#>  14.               └─pkgdown:::data_reference_topic(...)
#>  15.                 └─pkgdown:::run_examples(...)
#>  16.                   └─pkgdown:::highlight_examples(code, topic, env = env)
#>  17.                     └─downlit::evaluate_and_highlight(...)
#>  18.                       └─evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#>  19.                         ├─base::withRestarts(...)
#>  20.                         │ └─base (local) withRestartList(expr, restarts)
#>  21.                         │   ├─base (local) withOneRestart(withRestartList(expr, restarts[-nr]), restarts[[nr]])
#>  22.                         │   │ └─base (local) doWithOneRestart(return(expr), restart)
#>  23.                         │   └─base (local) withRestartList(expr, restarts[-nr])
#>  24.                         │     └─base (local) withOneRestart(expr, restarts[[1L]])
#>  25.                         │       └─base (local) doWithOneRestart(return(expr), restart)
#>  26.                         ├─evaluate:::with_handlers(...)
#>  27.                         │ ├─base::eval(call)
#>  28.                         │ │ └─base::eval(call)
#>  29.                         │ └─base::withCallingHandlers(...)
#>  30.                         ├─base::withVisible(eval(expr, envir))
#>  31.                         └─base::eval(expr, envir)
#>  32.                           └─base::eval(expr, envir)
#>  33.                             ├─base::try(f1(x = 1))
#>  34.                             │ └─base::tryCatch(...)
#>  35.                             │   └─base (local) tryCatchList(expr, classes, parentenv, handlers)
#>  36.                             │     └─base (local) tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>  37.                             │       └─base (local) doTryCatch(return(expr), name, parentenv, handler)
#>  38.                             └─f1(x = 1)
#>  39.                               └─ecokit::check_args(...)