Retrieve CHELSA Data Links

Fetches links to CHELSA climate data files from a specified base URL, filters them to include only links for *.tif files for variables available under current and future climate scenarios, and extracts metadata to create a tibble with detailed file information.

Usage

get_chelsa_links(base_url = "https://os.zhdk.cloud.switch.ch/chelsav2/")

Arguments

base_url

Base URL of the CHELSA repository. Defaults to "https://os.zhdk.cloud.switch.ch/chelsav2/".

Value

A tibble with the following columns:

url (character): Full URL of the data file.

relative_url (character): Relative URL, excluding the base URL.

file_name (character): Name of the data file.

dir_name (character): Directory path of the file.

climate_scenario (character): Climate scenario. Values are: "current", "ssp126", "ssp370", and "ssp585".

climate_model (character): Climate model. Values are: "Current", "GFDL-ESM4", "IPSL-CM6A-LR", "MPI-ESM1-2-HR", "MRI-ESM2-0", and "UKESM1-0-LL".

year (character): Year range. Values are "1981-2010", "2011-2040", "2041-2070", and "2071-2100".

var_name (character): Variable name, e.g., "bio1".

long_name (character): Full variable name, e.g., "mean annual air temperature".

unit (character): Measurement unit, e.g., "°C".

scale (numeric): Scale factor for the variable.

offset (numeric): Offset value for the variable.

explanation (character): Brief description of the variable.

Author

Ahmed El-Gabbas

Examples

library(tibble)
library(dplyr)
options(tibble.print_max = 200)

CHELSA_links <- get_chelsa_links()

dplyr::glimpse(CHELSA_links)
#> Rows: 2,116
#> Columns: 13
#> $ url              <chr> "https://os.zhdk.cloud.switch.ch/chelsav2/GLOBAL/clim…
#> $ relative_url     <chr> "GLOBAL/climatologies/1981-2010/bio/CHELSA_bio1_1981-…
#> $ file_name        <chr> "CHELSA_bio1_1981-2010_V.2.1.tif", "CHELSA_bio2_1981-…
#> $ dir_name         <chr> "GLOBAL/climatologies/1981-2010/bio", "GLOBAL/climato…
#> $ climate_scenario <chr> "current", "current", "current", "current", "current"…
#> $ climate_model    <chr> "current", "current", "current", "current", "current"…
#> $ year             <chr> "1981-2010", "1981-2010", "1981-2010", "1981-2010", "…
#> $ var_name         <chr> "bio1", "bio2", "bio3", "bio4", "bio5", "bio6", "bio7…
#> $ long_name        <chr> "mean annual air temperature", "mean diurnal air temp…
#> $ unit             <chr> "°C", "°C", "°C", "°C/100", "°C", "°C", "°C", "°C", "…
#> $ scale            <dbl> 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1…
#> $ offset           <dbl> -273.15, 0.00, 0.00, 0.00, -273.15, -273.15, 0.00, -2…
#> $ explanation      <chr> "mean annual daily mean air temperatures averaged ove…

# Count the number of files per climate scenario, model, and year
dplyr::count(CHELSA_links, climate_scenario, climate_model, year)
#> # A tibble: 46 × 4
#>    climate_scenario climate_model year          n
#>    <chr>            <chr>         <chr>     <int>
#>  1 current          current       1981-2010    46
#>  2 ssp126           GFDL-ESM4     2011-2040    46
#>  3 ssp126           GFDL-ESM4     2041-2070    46
#>  4 ssp126           GFDL-ESM4     2071-2100    46
#>  5 ssp126           IPSL-CM6A-LR  2011-2040    46
#>  6 ssp126           IPSL-CM6A-LR  2041-2070    46
#>  7 ssp126           IPSL-CM6A-LR  2071-2100    46
#>  8 ssp126           MPI-ESM1-2-HR 2011-2040    46
#>  9 ssp126           MPI-ESM1-2-HR 2041-2070    46
#> 10 ssp126           MPI-ESM1-2-HR 2071-2100    46
#> # ℹ 36 more rows

CHELSA_links %>%
  dplyr::count(var_name, long_name, unit, scale, offset, explanation)
#> # A tibble: 46 × 7
#>    var_name long_name                       unit  scale offset explanation     n
#>    <chr>    <chr>                           <chr> <dbl>  <dbl> <chr>       <int>
#>  1 bio1     mean annual air temperature     °C      0.1  -273. mean annua…    46
#>  2 bio10    mean daily mean air temperatur… °C      0.1  -273. The warmes…    46
#>  3 bio11    mean daily mean air temperatur… °C      0.1  -273. The coldes…    46
#>  4 bio12    annual precipitation amount     kg m…   0.1     0  Accumulate…    46
#>  5 bio13    precipitation amount of the we… kg m…   0.1     0  The precip…    46
#>  6 bio14    precipitation amount of the dr… kg m…   0.1     0  The precip…    46
#>  7 bio15    precipitation seasonality       kg m…   0.1     0  The Coeffi…    46
#>  8 bio16    mean monthly precipitation amo… kg m…   0.1     0  The wettes…    46
#>  9 bio17    mean monthly precipitation amo… kg m…   0.1     0  The driest…    46
#> 10 bio18    mean monthly precipitation amo… kg m…   0.1     0  The warmes…    46
#> # ℹ 36 more rows

print(CHELSA_links, n = 200)
#> # A tibble: 2,116 × 13
#>     url     relative_url file_name dir_name climate_scenario climate_model year 
#>     <chr>   <chr>        <chr>     <chr>    <chr>            <chr>         <chr>
#>   1 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… current          current       1981…
#>   2 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… current          current       1981…
#>   3 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… current          current       1981…
#>   4 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… current          current       1981…
#>   5 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… current          current       1981…
#>   6 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… current          current       1981…
#>   7 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… current          current       1981…
#>   8 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… current          current       1981…
#>   9 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… current          current       1981…
#>  10 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… current          current       1981…
#>  11 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… current          current       1981…
#>  12 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… current          current       1981…
#>  13 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… current          current       1981…
#>  14 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… current          current       1981…
#>  15 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… current          current       1981…
#>  16 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… current          current       1981…
#>  17 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… current          current       1981…
#>  18 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… current          current       1981…
#>  19 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… current          current       1981…
#>  20 https:… GLOBAL/clim… CHELSA_f… GLOBAL/… current          current       1981…
#>  21 https:… GLOBAL/clim… CHELSA_f… GLOBAL/… current          current       1981…
#>  22 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… current          current       1981…
#>  23 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… current          current       1981…
#>  24 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… current          current       1981…
#>  25 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… current          current       1981…
#>  26 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… current          current       1981…
#>  27 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… current          current       1981…
#>  28 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… current          current       1981…
#>  29 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… current          current       1981…
#>  30 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… current          current       1981…
#>  31 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… current          current       1981…
#>  32 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… current          current       1981…
#>  33 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… current          current       1981…
#>  34 https:… GLOBAL/clim… CHELSA_k… GLOBAL/… current          current       1981…
#>  35 https:… GLOBAL/clim… CHELSA_k… GLOBAL/… current          current       1981…
#>  36 https:… GLOBAL/clim… CHELSA_k… GLOBAL/… current          current       1981…
#>  37 https:… GLOBAL/clim… CHELSA_k… GLOBAL/… current          current       1981…
#>  38 https:… GLOBAL/clim… CHELSA_k… GLOBAL/… current          current       1981…
#>  39 https:… GLOBAL/clim… CHELSA_k… GLOBAL/… current          current       1981…
#>  40 https:… GLOBAL/clim… CHELSA_l… GLOBAL/… current          current       1981…
#>  41 https:… GLOBAL/clim… CHELSA_n… GLOBAL/… current          current       1981…
#>  42 https:… GLOBAL/clim… CHELSA_n… GLOBAL/… current          current       1981…
#>  43 https:… GLOBAL/clim… CHELSA_n… GLOBAL/… current          current       1981…
#>  44 https:… GLOBAL/clim… CHELSA_n… GLOBAL/… current          current       1981…
#>  45 https:… GLOBAL/clim… CHELSA_s… GLOBAL/… current          current       1981…
#>  46 https:… GLOBAL/clim… CHELSA_s… GLOBAL/… current          current       1981…
#>  47 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  48 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  49 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  50 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  51 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  52 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  53 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  54 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  55 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  56 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  57 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  58 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  59 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  60 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  61 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  62 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  63 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  64 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  65 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  66 https:… GLOBAL/clim… CHELSA_f… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  67 https:… GLOBAL/clim… CHELSA_f… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  68 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  69 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  70 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  71 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  72 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  73 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  74 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  75 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  76 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  77 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  78 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  79 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  80 https:… GLOBAL/clim… CHELSA_k… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  81 https:… GLOBAL/clim… CHELSA_k… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  82 https:… GLOBAL/clim… CHELSA_k… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  83 https:… GLOBAL/clim… CHELSA_k… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  84 https:… GLOBAL/clim… CHELSA_k… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  85 https:… GLOBAL/clim… CHELSA_k… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  86 https:… GLOBAL/clim… CHELSA_l… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  87 https:… GLOBAL/clim… CHELSA_n… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  88 https:… GLOBAL/clim… CHELSA_n… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  89 https:… GLOBAL/clim… CHELSA_n… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  90 https:… GLOBAL/clim… CHELSA_n… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  91 https:… GLOBAL/clim… CHELSA_s… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  92 https:… GLOBAL/clim… CHELSA_s… GLOBAL/… ssp126           GFDL-ESM4     2011…
#>  93 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2041…
#>  94 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2041…
#>  95 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2041…
#>  96 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2041…
#>  97 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2041…
#>  98 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2041…
#>  99 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 100 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 101 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 102 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 103 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 104 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 105 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 106 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 107 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 108 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 109 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 110 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 111 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 112 https:… GLOBAL/clim… CHELSA_f… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 113 https:… GLOBAL/clim… CHELSA_f… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 114 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 115 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 116 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 117 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 118 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 119 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 120 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 121 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 122 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 123 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 124 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 125 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 126 https:… GLOBAL/clim… CHELSA_k… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 127 https:… GLOBAL/clim… CHELSA_k… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 128 https:… GLOBAL/clim… CHELSA_k… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 129 https:… GLOBAL/clim… CHELSA_k… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 130 https:… GLOBAL/clim… CHELSA_k… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 131 https:… GLOBAL/clim… CHELSA_k… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 132 https:… GLOBAL/clim… CHELSA_l… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 133 https:… GLOBAL/clim… CHELSA_n… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 134 https:… GLOBAL/clim… CHELSA_n… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 135 https:… GLOBAL/clim… CHELSA_n… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 136 https:… GLOBAL/clim… CHELSA_n… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 137 https:… GLOBAL/clim… CHELSA_s… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 138 https:… GLOBAL/clim… CHELSA_s… GLOBAL/… ssp126           GFDL-ESM4     2041…
#> 139 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 140 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 141 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 142 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 143 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 144 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 145 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 146 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 147 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 148 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 149 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 150 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 151 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 152 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 153 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 154 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 155 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 156 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 157 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 158 https:… GLOBAL/clim… CHELSA_f… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 159 https:… GLOBAL/clim… CHELSA_f… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 160 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 161 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 162 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 163 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 164 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 165 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 166 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 167 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 168 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 169 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 170 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 171 https:… GLOBAL/clim… CHELSA_g… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 172 https:… GLOBAL/clim… CHELSA_k… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 173 https:… GLOBAL/clim… CHELSA_k… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 174 https:… GLOBAL/clim… CHELSA_k… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 175 https:… GLOBAL/clim… CHELSA_k… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 176 https:… GLOBAL/clim… CHELSA_k… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 177 https:… GLOBAL/clim… CHELSA_k… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 178 https:… GLOBAL/clim… CHELSA_l… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 179 https:… GLOBAL/clim… CHELSA_n… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 180 https:… GLOBAL/clim… CHELSA_n… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 181 https:… GLOBAL/clim… CHELSA_n… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 182 https:… GLOBAL/clim… CHELSA_n… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 183 https:… GLOBAL/clim… CHELSA_s… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 184 https:… GLOBAL/clim… CHELSA_s… GLOBAL/… ssp126           GFDL-ESM4     2071…
#> 185 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           IPSL-CM6A-LR  2011…
#> 186 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           IPSL-CM6A-LR  2011…
#> 187 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           IPSL-CM6A-LR  2011…
#> 188 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           IPSL-CM6A-LR  2011…
#> 189 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           IPSL-CM6A-LR  2011…
#> 190 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           IPSL-CM6A-LR  2011…
#> 191 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           IPSL-CM6A-LR  2011…
#> 192 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           IPSL-CM6A-LR  2011…
#> 193 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           IPSL-CM6A-LR  2011…
#> 194 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           IPSL-CM6A-LR  2011…
#> 195 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           IPSL-CM6A-LR  2011…
#> 196 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           IPSL-CM6A-LR  2011…
#> 197 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           IPSL-CM6A-LR  2011…
#> 198 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           IPSL-CM6A-LR  2011…
#> 199 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           IPSL-CM6A-LR  2011…
#> 200 https:… GLOBAL/clim… CHELSA_b… GLOBAL/… ssp126           IPSL-CM6A-LR  2011…
#> # ℹ 1,916 more rows
#> # ℹ 6 more variables: var_name <chr>, long_name <chr>, unit <chr>, scale <dbl>,
#> #   offset <dbl>, explanation <chr>