Load Values for SpatRaster Objects into Memory

Ensures a SpatRaster object has its values loaded into memory, processing inputs as a SpatRaster object, a file path to a raster, or a raster package object (RasterLayer, RasterStack, RasterBrick). Converts and unpacks inputs as needed.

Usage

set_raster_values(raster = NULL)

Arguments

raster

A SpatRaster object, a file path to a raster file (e.g., .tif, .nc), or a raster package object (RasterLayer, RasterStack, RasterBrick). Cannot be NULL or missing.

Value

The SpatRaster object with all values loaded into memory.

Details

The function handles various types of input:

If a file path is provided, it attempts to load the raster using terra::rast().

If the input is a packed SpatRaster, it unpacks the raster using terra::unwrap().

If the input is a raster object from the raster package, it is converted to SpatRaster.

Author

Ahmed El-Gabbas

Examples

load_packages(terra, raster)

f <- system.file("ex/elev.tif", package="terra")

# SpatRaster
r <- terra::rast(f)
terra::inMemory(r)
#> [1] FALSE
r_mem <- set_raster_values(r)
terra::inMemory(r_mem)
#> [1] TRUE

# PackedSpatRaster
r2 <- terra::wrap(terra::rast(f))
r2_mem <- set_raster_values(r2)
terra::inMemory(r2_mem)
#> [1] TRUE

# raster
r3 <- raster::raster(f)
raster::inMemory(r3)
#> [1] FALSE
r3_mem <- set_raster_values(r3)
terra::inMemory(r3_mem)
#> [1] TRUE